requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-core>=0.3.0
gradio>=4.0.0
websockets>=12.0
asyncio
python-dotenv
